Try and handle connection errors more carefully
Do not parse gateway messages if the connection got closed or errored
out (not sure if it's the right thing to do yet)
Discard empty gateway messages

Add a comment note about WSSTLS shutdown implementation seemingly being
buggy?

Seems like one of the ways SSL connection can be closed is short_read
error. For now just treat any error as an indication that we need to
reset the connection.